Moved scripting for quitting on button press to separate script
The OnQuitButtonClicked script is now separate from the canvas switcher. This makes more sense as this functionality is not to do with switching between menu UIs. Also, it requires none of the variables needed for switching, so can be separated easily.
