+++ Client/vr-client/Assets/NarupaIMD/UI/ButtonQuitter.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5e987651657542189a4f8925cd46e55a
+timeCreated: 1698940614
